fix: resolve GCC -Werror warnings (useless-cast, float-conversion)
- Remove redundant static_cast<uLong> in cache_manager.cpp zlib calls
- Add static_cast<double> for all long double -> double assignments in
  solar_analysis library (Vector3d uses long double internally):
  orbital_calculator.cpp, historical_analyzer.cpp, mission_analyzer.cpp,
  advanced_analyzer.cpp, data_models.cpp, analysis_engine.cpp
- These were pre-existing issues surfaced by the CMakeLists rebuild

Verified clean on both macOS (Clang) and Docker (Ubuntu 22.04 / GCC).
All 294 unit tests pass.

+++ lib/solar_analysis/src/advanced_analyzer.cpp
@@ -179,7 +179,7 @@ std::vector<size_t> AdvancedAnalyzer::classify_anomalies(const std::vector<Anoma
 
   for (size_t i = 0; i < anomalies.size(); ++i) {
     double normalized = (scores[i] - min_score) / range;
-    categories[i] = static_cast<size_t>(normalized * (num_categories - 1));
+    categories[i] = static_cast<size_t>(normalized * static_cast<double>(num_categories - 1));
   }
 
   return categories;
